feat: add `mpc-indexer-core` create, move state management and telemetry interfaces
