+++ chain-signatures/node/src/stream/ops.rs
@@ -5,9 +5,10 @@ use crate::respond_bidirectional::CompletedTx;
 use crate::rpc::{ContractStateWatcher, RpcChannel};
 use crate::sign_bidirectional::{SignBidirectionalEventExt, SignStatus};
 use anchor_lang::prelude::Pubkey;
+use mpc_indexer_core::ChainTelemetry;
 use mpc_primitives::{
-    BidirectionalTx, BidirectionalTxId, ChainTelemetry, ExecutionOutcome,
-    RespondBidirectionalEvent, SignId, SignKind, Signature, SignatureRespondedEvent,
+    BidirectionalTx, BidirectionalTxId, ExecutionOutcome, RespondBidirectionalEvent, SignId,
+    SignKind, Signature, SignatureRespondedEvent,
 };
 use tokio::sync::mpsc;
 
@@ -444,9 +445,8 @@ mod tests {
     use alloy::primitives::{Address, B256};
     use cait_sith::protocol::Participant;
     use k256::{ProjectivePoint, Scalar};
-    use mpc_primitives::{
-        RespondBidirectionalTx, SignArgs, SignBidirectionalEvent, SignKind, StateManager,
-    };
+    use mpc_indexer_core::StateManager;
+    use mpc_primitives::{RespondBidirectionalTx, SignArgs, SignBidirectionalEvent, SignKind};
     use near_primitives::types::AccountId;
     use solana_sdk::pubkey::Pubkey;
     use std::time::Duration;
