+++ chain-signatures/primitives/src/lib.rs
@@ -4,7 +4,6 @@ mod chain;
 mod crypto;
 mod events;
 mod requests;
-mod traits;
 
 pub use backlog::{Checkpoint, CheckpointDigest, ConsensusCheckpointDigest, PendingTx};
 pub use bidirectional::{
@@ -17,7 +16,6 @@ pub use crypto::{
 };
 pub use events::{ChainEvent, ExecutionOutcome, SignatureRespondedEvent};
 pub use requests::{IndexedSignRequest, SignKind};
-pub use traits::{ChainTelemetry, NoopChainTelemetry, StateManager};
 
 pub const LATEST_MPC_KEY_VERSION: u32 = 1;
 pub const LEGACY_MPC_KEY_VERSION_0: u32 = 0;
